More fixing mypy

+++ sleplet/functions/flm/noise_earth.py
@@ -8,7 +8,7 @@ from sleplet.utils.string_methods import convert_camel_case_to_snake_case, filen
 
 @dataclass(kw_only=True)
 class NoiseEarth(F_LM):
-    SNR: float = 10
+    SNR: int = 10
 
     def __post_init__(self) -> None:
         super().__post_init__()
